+++ seed/csharp-sdk/basic-auth-environment-variables/src/SeedBasicAuthEnvironmentVariables.Test/Unit/MockServer/BaseMockServerTest.cs
@@ -26,7 +26,7 @@ public class BaseMockServerTest
         );
 
         // Initialize the Client
-        Client = new SeedBasicAuthEnvironmentVariablesClient("USERNAME", "PASSWORD");
+        Client = new SeedBasicAuthEnvironmentVariablesClient();
 
         RequestOptions = new RequestOptions { BaseUrl = Server.Urls[0] };
     }