+++ seed/ts-sdk/code-samples/src/core/fetcher/stream-wrappers/UndiciStreamWrapper.ts
@@ -160,8 +160,12 @@ export class UndiciStreamWrapper<ReadFormat extends Uint8Array | Uint16Array | U
 
         while (true) {
             const { done, value } = await this.reader.read();
-            if (done) break;
-            if (value) chunks.push(value);
+            if (done) {
+                break;
+            }
+            if (value) {
+                chunks.push(value);
+            }
         }
 
         const decoder = new TextDecoder(this.encoding || "utf-8");